Refactor declaration
